+++ guillotina/db/transaction.py
@@ -1,12 +1,4 @@
-import asyncio
-import logging
-import sys
-import time
 from collections import OrderedDict
-from typing import AsyncIterator
-from typing import List
-from typing import Optional
-
 from guillotina import task_vars
 from guillotina._settings import app_settings
 from guillotina.component import get_adapter
@@ -23,7 +15,15 @@ from guillotina.exceptions import TIDConflictError
 from guillotina.exceptions import TransactionClosedException
 from guillotina.profile import profilable
 from guillotina.utils import lazy_apply
+from typing import AsyncIterator
+from typing import List
+from typing import Optional
 from zope.interface import implementer
+
+import asyncio
+import logging
+import sys
+import time
 
 
 _EMPTY = "__<EMPTY VALUE>__"
@@ -73,10 +73,10 @@ class Transaction:
     _skip_commit = False
 
     def __init__(self, manager, loop=None, read_only=False):
-        self.initialize(read_only)
-
         # Transaction Manager
         self._manager = manager
+
+        self.initialize(read_only)
 
         logger.debug("new transaction")
 
@@ -89,12 +89,6 @@ class Transaction:
         # this provides a lock for each transaction
         # which would correspond with one connection
         self._lock = asyncio.Lock(loop=loop)
-
-        self._strategy = get_adapter(self, ITransactionStrategy, name=manager._storage._transaction_strategy)
-
-        self._cache = query_adapter(self, ITransactionCache, name=app_settings["cache"]["strategy"])
-
-        self._query_count_start = self._query_count_end = 0
 
     def initialize(self, read_only):
         self._read_only = read_only
@@ -114,6 +108,12 @@ class Transaction:
 
         # List of (hook, args, kws) tuples added by addAfterCommitHook().
         self._after_commit = []
+
+        self._cache = query_adapter(self, ITransactionCache, name=app_settings["cache"]["strategy"])
+        self._strategy = get_adapter(
+            self, ITransactionStrategy, name=self._manager._storage._transaction_strategy
+        )
+        self._query_count_start = self._query_count_end = 0
 
     def get_query_count(self):
         """
@@ -268,6 +268,7 @@ class Transaction:
                 continue
             ob.__dict__[key] = value
         ob.__serial__ = new.__serial__
+        ob.__txn__ = self
 
     @cache(lambda oid: {"oid": oid}, True)
     async def _get(self, oid):
@@ -349,10 +350,6 @@ class Transaction:
 
     @profilable
     async def _store_object(self, obj, uid, added=False):
-        # Modified objects
-        if obj.__txn__ is not self and obj.__txn__ is not None:
-            raise Exception(f"Invalid reference to txn: {obj}")
-
         # There is no serial
         if added:
             serial = None
@@ -376,8 +373,6 @@ class Transaction:
         for oid, obj in self.modified.items():
             await self._store_object(obj, oid)
         for oid, obj in self.deleted.items():
-            if obj.__txn__ is not self and obj.__txn__ is not None:
-                raise Exception(f"Invalid reference to txn: {obj}")
             await self._manager._storage.delete(self, oid)
 
     @profilable